+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog ##
 
+### Unreleased
+
+#### Fixed
+
+- The contextmenu now checks `visible` on items once instead of twice. Fixes [issue 612](https://github.com/swisnl/jQuery-contextMenu/issues/612).
+
 ### 3.0.0-beta.1
 
 #### Migrating